Moving elements into elements.py Moving calculation into calculation.py moving system in system.py
